requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
gunicorn==21.2.0
uvloop==0.19.0
httptools==0.6.1
python-dotenv==1.0.0  # Added python-dotenv
pyyaml==6.0.1  # For YAML support
